Create Alert if user is disabled in AD
I manage a small app with about 60 users, all managed internally. Because I don’t get alerts when users become “inactive” (quit, terminated, etc.), I wrote the below script to ensure users are deleted when they leave. The script a list of user id’s from a text file and 3 AD groups then does a compare. If there is a change, meaning one of the users is no longer in one of the three specified groups; an email is sent out and I know to delete them from the system. Here is my current script: